% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_similarity_judgments.R
\name{determine_trial_id}
\alias{determine_trial_id}
\title{Label the trial id based on sound_x and sound_y of acoustic similarity trial.}
\usage{
determine_trial_id(frame)
}
\description{
Label the trial id based on sound_x and sound_y of acoustic similarity trial.
}
