% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-statistical-results.R
\name{glmer_mod_results}
\alias{glmer_mod_results}
\title{Present the results of a hierarchical generalized linear model as in a results section.}
\usage{
glmer_mod_results(glmer_mod, param)
}
\description{
Present the results of a hierarchical generalized linear model as in a results section.
}

