% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{label_outliers}
\alias{label_outliers}
\title{Label outlier subjects based on catch trial performance.}
\usage{
label_outliers(transcription_matches)
}
\description{
Label outlier subjects based on catch trial performance.
}

