% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{read_wordsintransition_chunks}
\alias{read_wordsintransition_chunks}
\title{Read all available knitr chunks in the wordsintransition package.}
\usage{
read_wordsintransition_chunks()
}
\description{
Read all available knitr chunks in the wordsintransition package.
}
