% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_sound_names.R
\name{recode_block_transition}
\alias{recode_block_transition}
\title{Recode the transition between blocks}
\usage{
recode_block_transition(frame)
}
\description{
Recode the transition between blocks
}
