% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_similarity_judgments.R
\name{recode_edge_generations}
\alias{recode_edge_generations}
\title{Recode edge generations for acoustic similarity judgments.}
\usage{
recode_edge_generations(frame)
}
\description{
Recode edge generations for acoustic similarity judgments.
}
