% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_sound_names.R
\name{recode_lsn_word_type}
\alias{recode_lsn_word_type}
\title{Recode the word type of the word learned in the LSN experiment.}
\usage{
recode_lsn_word_type(frame)
}
\description{
Recode the word type of the word learned in the LSN experiment.
}
